refactor: 결과를 관리하는 ResultController를 MainController에서 분리
